% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{distance}
\usage{
distance(X, medians)
}
\arguments{
\item{X}{a matrix, the dataset being clustered}

\item{medians}{a matrix, dedians of the clusters}
}
\value{
a matrix, distance between each point and each median
}
\description{
Calculates the Manhanttan distance between the medians and every point in the dataset
}
\examples{
A <- matrix(
c(1,2,3,4,5,6),
nrow = 3,
ncol = 2,
byrow = TRUE)
m <- matrix(
  c(1,1,2,2),
  nrow = 2,
  ncol = 2,
 byrow = TRUE)
distance(A,m)

}
