% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmedians_cluster.R
\name{kmedians}
\alias{kmedians}
\title{kmedians}
\usage{
kmedians(X, num_clusters, n_it = 100)
}
\arguments{
\item{X}{a matrix, the dataset being clustered}

\item{num_clusters}{integer, the desired number of clusters}

\item{n_it}{integer, number of iterations}
}
\value{
list, contains both medians and labes
        medians: matrix
        The coordinates of the medians for each cluster

        labels: list
         List that has the assignment of the cluster for each point in the dataset
}
\description{
Groups the points in your dataset ,X, into the desired number of clusters, based on the median distance between the points.
This function uses random intilization to assign the first medians and then will update the medians and
the group assignments until the assignment does not change.
}
\examples{




}
