% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmedians_summary.R
\name{summary}
\alias{summary}
\title{summary}
\usage{
summary(X, medians, labels)
}
\arguments{
\item{X}{matrix,the dataset being clustered}

\item{medians}{matrix, coordinates of each cluster median}

\item{labels}{list, array with the assignment of the cluster for each point in the dataset}
}
\value{
dataframe. Returns a dataframe with 6 columns and number of rows will be the number of clusters. The labels of the columns:
         Cluster labels, Median Coordinates, Number of Points in Cluster, Average Distance, Minimum Distance, Maximum Distance
}
\description{
Generates a table to display the cluster labels, the coordinates of the cluster medians,
number of points in each cluster, the average distance within the cluster,
the maximum distance within the cluster and the minimum distance within the cluster.
}
\examples{

}
